Added Quiz controller

+++ src/main/java/com/quizmasterpro/quizmaterpro/Services/QuizService.java
@@ -1,11 +1,8 @@
 package com.quizmasterpro.quizmaterpro.Services;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
@@ -17,6 +14,7 @@ import org.springframework.web.client.RestTemplate;
 import com.quizmasterpro.quizmaterpro.Dtos.Quiz.QuizAPIReq;
 import com.quizmasterpro.quizmaterpro.Dtos.Quiz.QuizAPIResp;
 import com.quizmasterpro.quizmaterpro.Dtos.Quiz.QuizCreateDto;
+import com.quizmasterpro.quizmaterpro.Dtos.Quiz.QuizSubmitDto;
 import com.quizmasterpro.quizmaterpro.Models.Quiz;
 import com.quizmasterpro.quizmaterpro.Models.Topic;
 import com.quizmasterpro.quizmaterpro.Models.User;
@@ -33,20 +31,19 @@ public class QuizService {
      @Autowired
     private TopicRepository topicRepository;
 
-    public List<Quiz> getAllQuizzes() {
-        return quizRepository.findAll();
+    public Quiz getById(String id) {
+        return quizRepository.findById(id).orElseThrow(()->new IllegalArgumentException("No quiz found"));
     }
 
-    public Optional<Quiz> getQuizById(UUID id) {
-        return quizRepository.findById(id);
-    }
-
-    public List<Quiz> getQuizzesByUserId(UUID userId) {
+    public List<Quiz> getByUserId(String userId) {
         return quizRepository.findByUserId(userId);
     }
 
     public Quiz create(QuizCreateDto quizCreateDto) {
         User user = userRepository.findById(quizCreateDto.getUserId()).orElseThrow(()->new UsernameNotFoundException("User not found"));
+        if(quizRepository.findByUserId(quizCreateDto.getUserId()).size()>=5){
+            throw new IllegalArgumentException("You can only give 5 quizzes in free plan");
+        }
         Topic topic =topicRepository.findById(quizCreateDto.getTopicId()).orElseThrow(()->new IllegalArgumentException("Topic not found"));
         HttpHeaders headers = new HttpHeaders();
         RestTemplate restTemplate = new RestTemplate();
@@ -56,7 +53,7 @@ public class QuizService {
         quizAPIReq.setNoOfQuestions(quizCreateDto.getTotalQuestions());
         quizAPIReq.setDifficulty(quizCreateDto.getDifficulty());
         HttpEntity<QuizAPIReq> request = new HttpEntity<>(quizAPIReq, headers);
-        ResponseEntity<QuizAPIResp> response = restTemplate.postForEntity("http://localhost:3000/gemini", request, QuizAPIResp.class);
+        ResponseEntity<QuizAPIResp> response = restTemplate.postForEntity("https://quiz-master-pro-gemini-layer.vercel.app/gemini", request, QuizAPIResp.class);
         QuizAPIResp resp = response.getBody();
 
         Quiz quiz = new Quiz();
@@ -72,7 +69,31 @@ public class QuizService {
         return quiz;
     }
 
-    public void deleteQuiz(UUID id) {
+    public Quiz submit(String quizId, QuizSubmitDto quizSubmitDto){
+        Quiz quiz = quizRepository.findById(quizId).orElseThrow(()->new IllegalArgumentException("No quiz found"));
+        userRepository.findById(quizSubmitDto.getUserId()).orElseThrow(()->new UsernameNotFoundException("User not found"));
+        if(quiz.isCompleted()){
+            throw new IllegalArgumentException("You already have completed the quiz");
+        }
+        quiz.setUserResponses(quizSubmitDto.getUserResponses());
+        List<String> answers = quiz.getAnswers();
+        List<String> userResponses = quizSubmitDto.getUserResponses();
+        int correctAnswers = 0;
+        for(int i=0;i<answers.size();i++){
+            if(answers.get(i).equals(userResponses.get(i))){
+                correctAnswers++;
+            }
+        }
+        quiz.setCorrectAnswers(correctAnswers);
+        quiz.setCompleted(true);
+        quizRepository.save(quiz);
+        return quiz;
+    }
+
+    public void delete(String id) {
+        if (quizRepository.findById(id).isEmpty()) {
+            throw new IllegalArgumentException("No quiz found to delete");
+        }
         quizRepository.deleteById(id);
     }
 }
